requirements.txt
python-telegram-bot>=21,<22
openai>=1.35,<2
python-dotenv>=1.0,<2
